Update notebook for cleaning BioKG from benchmarks
Added some details to the notebook, noting that removing benchmarks from
BioKG also removes entities from the graph entirely.
To be able to compare with methods like ComplEx in the benchmarks, we
thus now create *transductive* version of the benchmarks. The notebook
now takes care of this.
The notebook will now create an extra directory for storing BioKG
cleaned of benchmark related data, provisionally called BioKG-B.
The directory contains
- The BioKG links with no benchmark pairs
- The benchmarks, containing only entities in the cleaned version of the
graph