Subclass BioBLP from existing models
In the previous implementation, we implemented BioBLP as a set of
encoders for entities with descriptions, plus lookup tables for entities
with missing descriptions. To this we added the interaction function of
a particular model, like RotatE. However, to be able to fully reproduce
the results of RotatE without any kind of encoder, we had to ensure that
we were using the same initialization, and the normalization of relation
embeddings to unit norm vectors.
Instead of reimplementing these requirements (which change depending on
whether the interaction is RotatE, ComplEx, etc.), we now subclass
BioBLP from RotatE. In future work we will generalize this such that
BioBLP can be subclass of any other model dynamically.