Fix memory leak caused by forward hook
To hack into the forward pass of vanilla RotatE, the previous
implementation of BioBLP used a forward hook that is run by PyTorch
after calling a forward pass. It turns out that when the hooked method
has a reference to the model, this causes a memory leak:
https://github.com/pytorch/pytorch/issues/46386

In this commit we instead override the _plain_forward method of PyKEEN's
Embedding module, to encode entities when necessary, or get them from
the lookup table otherwise.